Make the build warning-clean and add a warnings-as-errors option
Cast the deliberate size_t wrap in the position iterator's initial value so
its signed/unsigned intent is explicit, and define _CRT_SECURE_NO_WARNINGS
for MSVC so the portable C99 fopen is used without a deprecation warning.
Add an opt-in STEGIFY_WERROR option that turns first-party warnings into
errors (/WX or -Werror), leaving third-party stb unaffected.

+++ src/stegify.c
@@ -145,7 +145,8 @@ extract_bit(uint8_t source)
 static void
 stegify_iter_init(stegify_position_iter_t *iter)
 {
-  iter->current_pos = -1;
+  /* Starts one before position 0; the first stegify_iter_next() wraps to 0. */
+  iter->current_pos = (size_t)-1;
   iter->step = 1;
 }
